allow to use binaries/strings in metadata keys
The problem is that when a binary or string is used in metadata keys
Json formatter treats such keys as values and mapping was wrong.

For example:

    meta = [
      {"a", 1},
      {"b", 2}
    ]

Expected mapping for such metadata is:

    %{"a": 1, "b": 2, "timestamp": ....}

But JsonFormatter treated such (non-atom) keys as values, so the
result was:

    %{"a": "a", "b": "b", "timestamp": ....}

This commit should fix this issue.